+++ biosys/apps/publish/views/export.py
@@ -6,7 +6,7 @@ from django.shortcuts import get_object_or_404
 from django.views.generic import View
 
 from main.models import Dataset, Record
-from main.utils_data_package import Exporter
+from main.api.exporters import DefaultExporter
 from main.utils_http import WorkbookResponse
 
 
@@ -14,7 +14,7 @@ class ExportDataSetView(View):
     def get(self, request, *args, **kwargs):
         ds = get_object_or_404(Dataset, pk=kwargs.get('pk'))
         qs = Record.objects.filter(dataset=ds).order_by('id')
-        exporter = Exporter(ds, qs)
+        exporter = DefaultExporter(ds, qs)
         wb = exporter.to_workbook()
         now = datetime.datetime.now()
         file_name = ds.name + '_' + now.strftime('%Y-%m-%d-%H%M%S') + '.xlsx'
@@ -25,7 +25,7 @@ class ExportDataSetView(View):
 class ExportTemplateView(View):
     def get(self, request, *args, **kwargs):
         ds = get_object_or_404(Dataset, pk=kwargs.get('pk'))
-        exporter = Exporter(ds)
+        exporter = DefaultExporter(ds)
         wb = exporter.to_workbook()
         file_name = ds.name + '_template'
         response = WorkbookResponse(wb, file_name)
