+++ biosys/settings.py
@@ -120,6 +120,7 @@ LOGIN_REDIRECT_URL = '/'
 AUTHENTICATION_BACKENDS = env('AUTHENTICATION_BACKENDS', [
     'django.contrib.auth.backends.ModelBackend',
 ])
+EXPORTER_CLASS = env('EXPORTER_CLASS', 'main.api.exporters.DefaultExporter')
 
 REST_FRAMEWORK = {
     'DEFAULT_AUTHENTICATION_CLASSES': env('REST_FRAMEWORK_DEFAULT_AUTHENTICATION_CLASSES', [
